moved get_workflow_action_buttons_html to utils.py instead
